requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.5.0
pyaudio>=0.2.13
pytest>=7.3.1
colorama>=0.4.6  # For better console output
tqdm>=4.65.0  # For progress bars